Implemented Ghost switching off Lights
Defined Function to switch off all light sources in the LightSwitchView. Added this function as listener to the OnLightsOffByGhostEvent action.

Added the the function to change player state on light off as listener to the action.

Defined function to play ghost giggle sound effect in the LightsOffByGhostEvent script and added the function as listener to the above action. Specified number of keys player needs in order to trigger the action. Invoked the above action when player enters collider of the game object to which this script is attached.