Created Rat Rush Action
Created RatRushEvent action as instance of EventContoller in EventService.

Added Camera Shake function from CameraShake script, red vignette function from GameUIView script and function to increase player insanity from PlayerSanity script as listeners to the RatRushEvent action.

Invoked RatRushEvent action in the RatRushEvent script when player enters the collider of the game object to which this script is attached.